Move pure sort logic to a file
